Moved delete to storage services
